Add files via upload
Description
 
Wearable health monitoring technology makes it possible to measure physiological indicators without interfering with daily activities and provides data that is indicative of a person's health and well-being. To acquire physiological data from the plantar aspect of the foot that may be evaluated to get insight into a
person's health, one such device can be placed in an insole, and the knee heals. An integrated understanding of a person's health and well-being can be achieved without interfering with daily activities by creating a system that can measure parameters such as PPD, gait characteristics, foot temperature, heart rate, and knee position. All the data are stored in the cloud. The proposed device can be used for a wide range of purposes, including the identification of pathologies, monitoring of medical conditions, and gait analysis. The data are processed using a machine learning - artificial neural network technique and stored in the cloud for later study. Real-time monitoring of the patient's schematic was done using the Unity software.


=> We first make sure the sensor is functioning effectively for our benefit before making the necessary alterations.
=> We must gather data from the sensors using the ESP-WROOM 32s microcontroller, manipulate the raw data, and transform it into real-time data.
=>The data transport to the website also used the same microcontroller. similar to https://friendly-dragon-0a92eb.netlify.app
=>We build the website using React JS and My SQL. which aids the physician's analysis.
=> We track the node using the ADXL335 Accelerometer and use the unity program to turn it into an animation. Animation is used to track the patent in a virtual environment.
=> The Force sensor, which helps to track the pressure of the leg at various locations, was mounted on the Feet.